renaming helpers to make it more generic
