+++ tee-worker/identity/ts-tests/integration-tests/common/utils/index.ts
@@ -13,4 +13,4 @@ export * from './integration-setup';
 export * from './storage';
 export * from './vc-helper';
 export * from './type-creators';
-export * from './omni-account-helpers';
+export * from './native-request-helpers';
